feat(pydantic): fix tests, add release notes
