+++ RELEASE.MD
@@ -1,0 +1,32 @@
+Adds support for converting pydantic conlist.
+Note that constraint is not enforced in the graphql type.
+Thus, we recommend always working on the pydantic type such that the validated is enforced.
+
+```python
+from pydantic import BaseModel, conlist
+import strawberry
+
+class Example(BaseModel):
+    friends: conlist(str, min_items=1)
+
+@strawberry.experimental.pydantic.type(model=Example, all_fields=True, is_input=True)
+class ExampleGQL:
+    ...
+
+@strawberry.type
+class Query:
+    @strawberry.field()
+    def test(self, example: ExampleGQL) -> None:
+        # if to_pydantic() is not called, there will be no validation that
+        # friends has at least one item
+        print(example.to_pydantic())
+
+schema = strawberry.Schema(query=Query)
+```
+
+The converted graphql type is
+```
+input ExampleGQL {
+  friends: [String!]!
+}
+```
