+++ strawberry/experimental/pydantic/fields.py
@@ -1,5 +1,5 @@
 from decimal import Decimal
-from typing import List, Optional
+from typing import Any, List, Optional, Type
 from uuid import UUID
 
 import pydantic
@@ -60,20 +60,21 @@ FIELDS_MAP = {
 }
 
 
-def get_basic_type(type_):
+def get_basic_type(type_) -> Type[Any]:
     if isinstance(type_, type):
         if issubclass(type_, pydantic.ConstrainedInt):
             return int
         if issubclass(type_, pydantic.ConstrainedStr):
             return str
 
-    if "ConstrainedListValue" in type_.__name__:
-        # pydantic ConstrainedListValue is created dynamically
-        # it isn't wrapped in a list, so we need to do it ourselves
-        return List[type_.item_type]
-    if "ConstrainedSetValue" in type_.__name__:
-        # sets don't exist in graphql
-        return List[type_.item_type]
+    if hasattr(type_, "__name__"):
+        if "ConstrainedListValue" in type_.__name__:
+            # pydantic ConstrainedListValue is created dynamically
+            # it isn't wrapped in a list, so we need to do it ourselves
+            return List[get_basic_type(type_.item_type)]  # type: ignore
+        if "ConstrainedSetValue" in type_.__name__:
+            # sets don't exist in graphql
+            return List[get_basic_type(type_.item_type)]  # type: ignore
 
     if type_ in FIELDS_MAP:
         type_ = FIELDS_MAP.get(type_)
